refactor: created a folder for each component
