+++ source/js/components/Confirmation/index.ts
@@ -1,0 +1,1 @@
+export * from './Confirmation';
