+++ source/js/components/GridElement/GridElement.tsx
@@ -1,0 +1,7 @@
+interface GridElementProps {
+  children: React.ReactChild | React.ReactChild[];
+  width: number;
+}
+export const GridElement = ({ children, width }: GridElementProps) => (
+  <div className={`o-grid-${width}@md`}>{children}</div>
+);
