+++ source/js/components/GridElement/index.ts
@@ -1,0 +1,1 @@
+export * from './GridElement';
