+++ source/js/components/GridRow/index.ts
@@ -1,0 +1,1 @@
+export * from './GridRow';
