+++ source/js/components/Loader/index.ts
@@ -1,0 +1,1 @@
+export * from './Loader';
