+++ source/js/components/Notice/index.ts
@@ -1,0 +1,1 @@
+export * from './Notice';
