+++ source/js/components/TextField/TextField.tsx
@@ -7,7 +7,7 @@ interface TextFieldInterface {
   type: HTMLInputTypeAttribute;
   required?: boolean;
 }
-const TextField = ({ onChange, value, id, label, type, required }: TextFieldInterface) => (
+export const TextField = ({ onChange, value, id, label, type, required }: TextFieldInterface) => (
   <div className="form-group">
     <label className="c-field__text--label">{label}</label>
     <div className="c-field c-field--text c-field--md c-field--radius-md">
@@ -18,5 +18,3 @@ const TextField = ({ onChange, value, id, label, type, required }: TextFieldInte
     </div>
   </div>
 );
-
-export default TextField;
