+++ source/js/components/index.ts
@@ -1,11 +1,10 @@
-export { default as DatePicker } from './DatePicker';
-export { default as Button } from './Button';
-export { default as TextField } from './TextField';
-export { default as Notice } from './Notice';
-export { default as Form } from './Form';
-export { default as Confirmation } from './Confirmation';
-export { default as GridRow } from './GridRow';
-export { default as GridElement } from './GridElement';
-export { default as ErrorList } from './ErrorList';
-export { default as Loader } from './Loader';
-export type { ConfirmationInterface } from './Confirmation';
+export * from './DatePicker';
+export * from './Button';
+export * from './TextField';
+export * from './Notice';
+export * from './Form';
+export * from './Confirmation';
+export * from './GridRow';
+export * from './GridElement';
+export * from './ErrorList';
+export * from './Loader';
